HumanAI/pyproject.toml
[project]
name = "humanai"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "livekit-agents[bey,google,openai,silero]~=1.3",
    "livekit-plugins-google>=1.3.10",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.1",
]
